envoy: extract Caches struct with delete logic
Separate it from the code that generates the Envoy config from
Ingresses.

That Caches struct needs to have some specific logic to, for example,
delete all the information associated with an ingress.
